return bad request when an error occur while transcribing

+++ RC-SpeechToText/Controllers/TranscriptionController.cs
@@ -52,10 +52,21 @@ namespace RC_SpeechToText.Controllers
             Converter converter = new Converter();
             // Call converter to convert the file to mono and bring back its file path. 
             string convertedFileLocation = converter.FileToWav(filePath);
+
+			if(convertedFileLocation == null)
+			{
+				return BadRequest("An error occured when converting the file to Wav");
+			}
+
             _logger.LogInformation(DateTime.Now.ToString(_dateConfig) + " - "+ this.GetType().Name +" \n Audio file " + audioFile.FileName + " converted to wav at " + convertedFileLocation);
 
             // Call the method that will get the transcription
             GoogleResult result = TranscriptionService.GoogleSpeechToText(convertedFileLocation);
+
+			if(result.Error != null)
+			{
+				return BadRequest("An error occured when transcribing the speech to text: " + result.Error);
+			}
 
             // Delete the converted file
             converter.DeleteFile(convertedFileLocation);
@@ -64,10 +75,20 @@ namespace RC_SpeechToText.Controllers
             // Create thumbnail
             var thumbnailPath = Directory.GetCurrentDirectory() + @"\wwwroot\assets\Thumbnails\";
             Directory.CreateDirectory(thumbnailPath);
-            converter.CreateThumbnail(filePath, thumbnailPath + audioFile.FileName + ".jpg");
+            var thumbnailImage = converter.CreateThumbnail(filePath, thumbnailPath + audioFile.FileName + ".jpg");
+
+			if(thumbnailImage == null)
+			{
+				return BadRequest("An error occured when creating the thumbnail");
+			}
 
             // Get user id by email
             var user = await _context.User.Where(u => u.Email == userEmail).FirstOrDefaultAsync();
+
+			if(user == null)
+			{
+				return BadRequest("An error occured when retrieving the user");
+			}
 
             // Create file
             //TODO: get the type of the object, if it is a Video or an Audio file 
@@ -107,7 +128,6 @@ namespace RC_SpeechToText.Controllers
 
                 await _context.Word.AddAsync(word);
                 await _context.SaveChangesAsync();
-
             }
 
             _logger.LogInformation(DateTime.Now.ToString(_dateConfig) + " - "+ this.GetType().Name +" \n Added file with title: " + file.Title + " to the database");
